requirements.txt
-f /usr/share/pip-wheels
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.5
django-crispy-forms==1.8.1
django-extensions==2.2.5
django-filter==2.2.0
django-taggit==1.3.0
djangorestframework==3.11.2
idna==3.3
Markdown==3.1.1
mysql==0.0.3
mysqlclient==2.1.0
oauthlib==3.1.1
pycparser==2.20
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.4.2
urllib3==1.26.7
